refactor(streams): extract ready-frame coalescing internals
- isolate ready-frame polling and coalescing in a private module
- move coalescing tests into a dedicated test module
- preserve byte ownership, error ordering, and stream lifecycle behavior
